Fix smooth sky color and implementation of bless

+++ src/main/java/cn/teampancake/theaurorian/client/renderer/level/TASkyRenderer.java
@@ -112,14 +112,14 @@ public class TASkyRenderer {
 
     public static Vec3 getSkyColor(ClientLevel level, Vec3 pos) {
         float timeOfDay = level.dimensionType().timeOfDay(1000L);
-        int rgbColor = smoothColorTransition((level.dayTime() - 6000L) / 24000);
+        int rgbColor = smoothColorTransition((level.dayTime() - 6000L) / 24000F);
         Vec3 vec3 = pos.subtract(2.0D, 2.0D, 2.0D).scale(0.25D);
         Vec3 vec31 = CubicSampler.gaussianSampleVec3(vec3, (x, y, z) -> Vec3.fromRGB24(rgbColor));
         float f1 = Mth.cos(timeOfDay * ((float) Math.PI * 2F)) * 2.0F + 0.5F;
         return new Vec3((float) vec31.x * f1, (float) vec31.y * f1, (float) vec31.z * f1);
     }
 
-    public static int smoothColorTransition(long t) {
+    public static int smoothColorTransition(float t) {
         Color currentColor = new Color(0x010e34);
         Color targetColor = new Color(getDaySkyColors().get(currentPhase));
         double d = Math.sin(2.0F * Math.PI * t + 0.25F * Math.PI);
